working notification on new messages
